refactor: Move modules/machinery/components/Count.tsx -> MachineryCount.tsx
