+++ frontend/src/modules/machinery/components/MachineryCount.tsx
